Testing publishing/saving of changes in personal branch
I want to make sure ahead of the start of class that students will be able to push changes to their own branches in this project, so their work is properly saved. Not sure if there's a better way I should do this for the demo project.

For this commit, I'm changing Unity version from 6000.0.51f1 to 6000.0.52f1 (the most recent one)